Revert moved functions
